+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Fix issue with removeView Marionette.Region method
+- Upgrade dependencies
 
 ## 2.0.0
 - Update README.md